requirements.txt
crewai>=0.28.0
crewai-tools[mcp]>=0.2.0
litellm>=1.30.0
mcp>=0.9.0
python-dotenv>=1.0.0
pillow>=10.0.0

